requirements.txt
aider-chat>=0.50.0
python-telegram-bot>=21.0
